requirements.txt
Flask==3.1.0
requests==2.32.3
python-dotenv==1.0.1
Pillow==11.1.0
